# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Examination Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: type
          in: query
          schema:
            type: string
            enum: [mcq]
        - name: topic
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Question list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionListResponse"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Exam Taking ──
...
components:
  schemas:
...
    Question:
      type: object
      properties:
        id:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
          enum: [mcq]
        subject:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        options:
          type: ["array", "null"]
          items:
            type: string
        correctAnswer:
          type: ["string", "null"]
        marks:
          type: integer
        markingScheme:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, questionText, questionType, subject, topic, difficulty, marks, createdAt]
    CreateQuestionBody:
      type: object
      properties:
        questionText:
          type: string
        questionType:
          type: string
          enum: [mcq]
        subject:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        marks:
          type: integer
        markingScheme:
          type: string
      required: [questionText, questionType, subject, topic, difficulty, marks]
    UpdateQuestionBody:
      type: object
      properties:
        questionText:
          type: string
        questionType:
          type: string
          enum: [mcq]
        subject:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        marks:
          type: integer
        markingScheme:
          type: string
    QuestionListResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [questions, total, page, totalPages]
...